refactor: separate ChatList
